+++ components/react/src/Particles.tsx
@@ -47,17 +47,19 @@ export default class Particles extends Component<IParticlesProps, IParticlesStat
     }
 
     forceUpdate(): void {
-        this.refresh();
-
-        super.forceUpdate();
+        this.refresh().then(() => {
+            super.forceUpdate();
+        });
     }
 
     componentDidMount(): void {
-        if (this.props.init) {
-            this.props.init(tsParticles);
-        }
+        (async () => {
+            if (this.props.init) {
+                await this.props.init(tsParticles);
+            }
 
-        this.loadParticles();
+            await this.loadParticles();
+        })();
     }
 
     componentWillUnmount(): void {
@@ -81,14 +83,14 @@ export default class Particles extends Component<IParticlesProps, IParticlesStat
         );
     }
 
-    private refresh(): void {
+    private async refresh(): Promise<void> {
         this.destroy();
 
-        this.loadParticles();
+        await this.loadParticles();
     }
 
-    private loadParticles(): void {
-        const cb = (container?: Container) => {
+    private async loadParticles(): Promise<void> {
+        const cb = async (container?: Container) => {
             if (this.props.container) {
                 (this.props.container as MutableRefObject<Container>).current = container;
             }
@@ -98,14 +100,14 @@ export default class Particles extends Component<IParticlesProps, IParticlesStat
             });
 
             if (this.props.loaded) {
-                this.props.loaded(container);
+                await this.props.loaded(container);
             }
         };
 
-        if (this.props.url) {
-            tsParticles.loadJSON(this.props.id, this.props.url).then(cb);
-        } else {
-            tsParticles.load(this.props.id, this.props.params ?? this.props.options).then(cb);
-        }
+        const container = this.props.url
+            ? await tsParticles.loadJSON(this.props.id, this.props.url)
+            : await tsParticles.load(this.props.id, this.props.params ?? this.props.options);
+
+        await cb(container);
     }
 }