+++ engine/src/Options/Classes/Options.ts
@@ -4,7 +4,7 @@ import { ParticlesOptions } from "./Particles/ParticlesOptions";
 import { BackgroundMask } from "./BackgroundMask/BackgroundMask";
 import type { RangeValue, RecursivePartial } from "../../Types";
 import { Background } from "./Background/Background";
-import { Plugins } from "../../Utils";
+import { deepExtend, Plugins } from "../../Utils";
 import type { IOptionLoader } from "../Interfaces/IOptionLoader";
 import { Theme } from "./Theme/Theme";
 import { ResponsiveMode, ThemeMode } from "../../Enums";
@@ -78,6 +78,7 @@ export class Options implements IOptions, IOptionLoader<IOptions> {
     pauseOnBlur;
     pauseOnOutsideViewport;
     preset?: string | string[];
+    style: RecursivePartial<CSSStyleDeclaration>;
     responsive: Responsive[];
     themes: Theme[];
     zLayers;
@@ -101,6 +102,7 @@ export class Options implements IOptions, IOptionLoader<IOptions> {
         this.pauseOnBlur = true;
         this.pauseOnOutsideViewport = true;
         this.responsive = [];
+        this.style = {};
         this.themes = [];
         this.zLayers = 100;
     }
@@ -181,6 +183,8 @@ export class Options implements IOptions, IOptionLoader<IOptions> {
 
         this.motion.load(data.motion);
         this.particles.load(data.particles);
+
+        this.style = deepExtend(this.style, data.style) as RecursivePartial<CSSStyleDeclaration>;
 
         Plugins.loadOptions(this, data);