+++ components/preact/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.38.0](https://github.com/matteobruni/tsparticles/compare/preact-particles@1.37.6...preact-particles@1.38.0) (2021-12-29)
+
+**Note:** Version bump only for package preact-particles
+
+
+
+
+
 ## [1.37.6](https://github.com/matteobruni/tsparticles/compare/preact-particles@1.37.5...preact-particles@1.37.6) (2021-12-24)