+++ demo/vue3/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.21.0](https://github.com/matteobruni/tsparticles/compare/particles.vue3-demo@1.20.6...particles.vue3-demo@1.21.0) (2021-12-29)
+
+**Note:** Version bump only for package particles.vue3-demo
+
+
+
+
+
 ## [1.20.6](https://github.com/matteobruni/tsparticles/compare/particles.vue3-demo@1.20.5...particles.vue3-demo@1.20.6) (2021-12-24)
 
 **Note:** Version bump only for package particles.vue3-demo