+++ paths/simplexNoise/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.38.0](https://github.com/matteobruni/tsparticles/compare/tsparticles-path-simplex-noise@1.37.6...tsparticles-path-simplex-noise@1.38.0) (2021-12-29)
+
+**Note:** Version bump only for package tsparticles-path-simplex-noise
+
+
+
+
+
 ## [1.37.6](https://github.com/matteobruni/tsparticles/compare/tsparticles-path-simplex-noise@1.37.5...tsparticles-path-simplex-noise@1.37.6) (2021-12-24)
 
 **Note:** Version bump only for package tsparticles-path-simplex-noise